Arise deprecation level, fix #353
