+++ mirai-core/src/commonMain/kotlin/net.mamoe.mirai/message/data/impl.kt
@@ -13,15 +13,77 @@
 
 package net.mamoe.mirai.message.data
 
+import net.mamoe.mirai.Bot
+import net.mamoe.mirai.utils.ExternalImage
 import net.mamoe.mirai.utils.MiraiExperimentalAPI
+import net.mamoe.mirai.utils.SinceMirai
 import kotlin.jvm.JvmMultifileClass
 import kotlin.jvm.JvmName
 import kotlin.jvm.JvmSynthetic
 import kotlin.native.concurrent.SharedImmutable
 
+// region image
+
+/**
+ * 所有 [Image] 实现的基类.
+ */
+internal abstract class AbstractImage : Image { // make sealed in 1.3.0 ?
+    @Deprecated(
+        """
+        不要自行实现 Image, 它必须由协议模块实现, 否则会无法发送也无法解析.
+    """, level = DeprecationLevel.HIDDEN
+    )
+    @Suppress("PropertyName", "DeprecatedCallableAddReplaceWith")
+    @get:JvmSynthetic
+    final override val DoNotImplementThisClass: Nothing?
+        get() = error("stub")
+
+    private var _stringValue: String? = null
+        get() = field ?: kotlin.run {
+            field = "[mirai:image:$imageId]"
+            field
+        }
+
+    final override fun toString(): String = _stringValue!!
+    final override fun contentToString(): String = "[图片]"
+}
+
+internal interface ConstOriginUrlAware : Image {
+    val originUrl: String
+}
+
+internal interface DeferredOriginUrlAware : Image {
+    fun getUrl(bot: Bot): String
+}
+
+internal interface SuspendDeferredOriginUrlAware : Image {
+    suspend fun getUrl(bot: Bot): String
+}
+
+/**
+ * 由 [ExternalImage] 委托的 [Image] 类型.
+ */
+@SinceMirai("1.1.0")
+@MiraiExperimentalAPI("Will be renamed to OfflineImage on 1.2.0")
+@Suppress("DEPRECATION_ERROR")
+internal class ExperimentalDeferredImage internal constructor(
+    @Suppress("CanBeParameter") private val externalImage: ExternalImage // for future use
+) : AbstractImage(), SuspendDeferredOriginUrlAware {
+    override suspend fun getUrl(bot: Bot): String {
+        TODO()
+    }
+
+    override val imageId: String = externalImage.calculateImageResourceId()
+}
+
+internal val firstOnlineBotInstance: Bot get() = Bot.botInstancesSequence.firstOrNull() ?: error("No Bot available")
+
+// endergion
+
 /////////////////////////
 //// IMPLEMENTATIONS ////
 /////////////////////////
+
 private fun Message.hasDuplicationOfConstrain(key: Message.Key<*>): Boolean {
     return when (this) {
         is SingleMessage -> (this as? ConstrainSingle<*>)?.key == key
